Rename IntersectionServer to intersection_server
